docs/datasheet/scripts/requirements.txt
pymupdf>=1.26.0
pdfminer.six>=20221105
pytesseract>=0.3.10
